Create AvatarImageView to handle user avatars
