+++ wallet/res/layout/contact_row.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="70dp">
 
-    <ImageView
+    <de.schildbach.wallet.ui.dashpay.AvatarImageView
         android:id="@+id/avatar"
         android:transitionName="avatar"
         android:layout_width="36dp"
