+++ wallet/src/de/schildbach/wallet/ui/HeaderBalanceFragment.java
@@ -55,6 +55,7 @@ import de.schildbach.wallet.data.BlockchainIdentityData;
 import de.schildbach.wallet.data.BlockchainState;
 import de.schildbach.wallet.rates.ExchangeRate;
 import de.schildbach.wallet.rates.ExchangeRatesViewModel;
+import de.schildbach.wallet.ui.dashpay.AvatarImageView;
 import de.schildbach.wallet_test.R;
 
 public final class HeaderBalanceFragment extends Fragment {
@@ -72,6 +73,7 @@ public final class HeaderBalanceFragment extends Fragment {
     private View view;
     private CurrencyTextView viewBalanceDash;
     private CurrencyTextView viewBalanceLocal;
+    private AvatarImageView dashPayUserAvatar;
 
     private boolean showLocalBalance;
 
@@ -131,6 +133,8 @@ public final class HeaderBalanceFragment extends Fragment {
         viewBalanceLocal.setInsignificantRelativeSize(1);
         viewBalanceLocal.setStrikeThru(!Constants.IS_PROD_BUILD);
 
+        dashPayUserAvatar = view.findViewById(R.id.dashpay_user_avatar);
+
         view.setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -152,10 +156,10 @@ public final class HeaderBalanceFragment extends Fragment {
             public void onChanged(BlockchainIdentityBaseData blockchainIdentityData) {
                 if (blockchainIdentityData != null
                         && blockchainIdentityData.getCreationState().ordinal() >= BlockchainIdentityData.CreationState.DONE.ordinal()) {
-                    String firstLetter = blockchainIdentityData.getUsername().substring(0, 1);
-                    setDefaultUserAvatar(firstLetter.toUpperCase());
+                    dashPayUserAvatar.setDefaultUserAvatar(blockchainIdentityData.getUsername().toUpperCase());
+                    dashPayUserAvatar.setVisibility(View.VISIBLE);
                 } else {
-                    setDefaultUserAvatar(null);
+                    dashPayUserAvatar.setVisibility(View.GONE);
                 }
             }
         });
@@ -191,33 +195,6 @@ public final class HeaderBalanceFragment extends Fragment {
 
         autoLockHandler.removeCallbacksAndMessages(null);
         super.onPause();
-    }
-
-    private void setDefaultUserAvatar(String letters) {
-        ImageView dashpayUserAvatar = view.findViewById(R.id.dashpay_user_avatar);
-        if (letters == null) {
-            // there is no username, so hide the image
-            dashpayUserAvatar.setVisibility(View.GONE);
-            return;
-        }
-        dashpayUserAvatar.setVisibility(View.VISIBLE);
-        float[] hsv = new float[3];
-        //Ascii codes for A: 65 - Z: 90, 0: 48 - 9: 57
-        float firstChar = letters.charAt(0);
-        float charIndex;
-        if (firstChar <= 57) { //57 == '9' in Ascii table
-            charIndex = (firstChar - 48f) / 36f; // 48 == '0', 36 == total count of supported
-        } else {
-            charIndex = (firstChar - 65f + 10f) / 36f; // 65 == 'A', 10 == count of digits
-        }
-        hsv[0] = charIndex * 360f;
-        hsv[1] = 0.3f;
-        hsv[2] = 0.6f;
-        int bgColor = Color.HSVToColor(hsv);
-        final TextDrawable defaultAvatar = TextDrawable.builder().beginConfig().textColor(Color.WHITE)
-                .useFont(ResourcesCompat.getFont(getContext(), R.font.montserrat_regular))
-                .endConfig().buildRound(letters, bgColor);
-        dashpayUserAvatar.setBackground(defaultAvatar);
     }
 
     private void updateView() {
