+++ wallet/src/de/schildbach/wallet/ui/dashpay/ContactSearchResultsAdapter.kt
@@ -17,18 +17,13 @@
 
 package de.schildbach.wallet.ui.dashpay
 
-import android.graphics.Color
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.AdapterView
 import android.widget.ArrayAdapter
-import android.widget.ImageView
 import android.widget.TextView
-import androidx.core.content.res.ResourcesCompat
 import androidx.recyclerview.widget.RecyclerView
-import com.amulyakhare.textdrawable.TextDrawable
-import com.bumptech.glide.Glide
 import de.schildbach.wallet.data.UsernameSearchResult
 import de.schildbach.wallet.data.UsernameSortOrderBy
 import de.schildbach.wallet_test.R
@@ -117,7 +112,7 @@ class ContactSearchResultsAdapter(private val onSortOrderChangedListener: OnSort
     open inner class ViewHolder(resId: Int, inflater: LayoutInflater, parent: ViewGroup) :
             RecyclerView.ViewHolder(inflater.inflate(resId, parent, false)) {
 
-        private val avatar by lazy { itemView.findViewById<ImageView>(R.id.avatar) }
+        private val avatar by lazy { itemView.findViewById<AvatarImageView>(R.id.avatar) }
         private val username by lazy { itemView.findViewById<TextView>(R.id.username) }
         private val displayName by lazy { itemView.findViewById<TextView>(R.id.displayName) }
 
@@ -132,10 +127,9 @@ class ContactSearchResultsAdapter(private val onSortOrderChangedListener: OnSort
             }
 
             if(dashPayProfile.avatarUrl.isNotEmpty()) {
-                Glide.with(avatar).load(dashPayProfile.avatarUrl).circleCrop()
-                        .placeholder(R.drawable.user5).into(avatar)
+                avatar.setUrl(dashPayProfile.avatarUrl)
             } else {
-                setDefaultUserAvatar(dashPayProfile.username.toUpperCase())
+                avatar.setDefaultUserAvatar(dashPayProfile.username.toUpperCase())
             }
 
             itemClickListener?.let { l ->
@@ -143,29 +137,6 @@ class ContactSearchResultsAdapter(private val onSortOrderChangedListener: OnSort
                     l.onItemClicked(it, usernameSearchResult)
                 }
             }
-        }
-
-        // TODO: how do we refactor this, the code is in three places
-        private fun setDefaultUserAvatar(letters: String) {
-            val dashpayUserAvatar: ImageView = itemView.findViewById(R.id.avatar)
-            dashpayUserAvatar.visibility = View.VISIBLE
-            val hsv = FloatArray(3)
-            //Ascii codes for A: 65 - Z: 90, 0: 48 - 9: 57
-            val firstChar = letters[0].toFloat()
-            val charIndex: Float
-            charIndex = if (firstChar <= 57) { //57 == '9' in Ascii table
-                (firstChar - 48f) / 36f // 48 == '0', 36 == total count of supported
-            } else {
-                (firstChar - 65f + 10f) / 36f // 65 == 'A', 10 == count of digits
-            }
-            hsv[0] = charIndex * 360f
-            hsv[1] = 0.3f
-            hsv[2] = 0.6f
-            val bgColor = Color.HSVToColor(hsv)
-            val defaultAvatar = TextDrawable.builder().beginConfig().textColor(Color.WHITE)
-                    .useFont(ResourcesCompat.getFont(itemView.context, R.font.montserrat_regular))
-                    .endConfig().buildRound(letters[0].toString(), bgColor)
-            dashpayUserAvatar.background = defaultAvatar
         }
     }
